refactor: change judgment item domain policy
judgment item can't be initialized with empty values such as 0
